move methods to utilities.py
